// frontend/src/app/features/business-partners/suppliers/models/supplier-update-request.model.ts
export interface SupplierUpdateRequest {
  code: string;
  companyName: string;

  contactName: string | null;
  email: string | null;
  phoneNumber: string | null;

  address: string;
  city: string;
  postalCode: string | null;
  country: string;

  vatNumber: string | null;

  active: boolean;
}